Cleanup and reorganize all the helm related code
