+++ frontend/packages/dev-console/src/components/helm/__tests__/HelmReleaseDetailsPage.spec.tsx
@@ -1,82 +1,34 @@
 import * as React from 'react';
-import { shallow } from 'enzyme';
-import { ErrorPage404 } from '@console/internal/components/error';
-import { DetailsPage } from '@console/internal/components/factory';
-import { LoadingBox, StatusBox } from '@console/internal/components/utils';
-import HelmReleaseDetailsPage, { LoadedHelmReleaseDetailsPage } from '../HelmReleaseDetailsPage';
-import { mockHelmReleases } from './helm-release-mock-data';
+import { shallow, ShallowWrapper } from 'enzyme';
+import NamespacedPage from '../../NamespacedPage';
+import HelmReleaseDetails from '../details/HelmReleaseDetails';
+import HelmReleaseDetailsPage from '../HelmReleaseDetailsPage';
 
-let helmReleaseDetailsPageProps: React.ComponentProps<typeof HelmReleaseDetailsPage>;
-let loadedHelmReleaseDetailsPageProps: React.ComponentProps<typeof LoadedHelmReleaseDetailsPage>;
+type Component = typeof HelmReleaseDetailsPage;
+type Props = React.ComponentProps<Component>;
+let helmReleaseDetailsProps: Props;
+let helmReleaseDetailsPage: ShallowWrapper<Props>;
 
 describe('HelmReleaseDetailsPage', () => {
   beforeEach(() => {
-    helmReleaseDetailsPageProps = {
-      secret: {
-        data: [
-          {
-            metadata: {
-              name: 'secret-name',
-              namespace: 'xyz',
-              creationTimestamp: '2020-01-13T05:42:19Z',
-              labels: {
-                name: 'helm-mysql',
-                owner: 'helm',
-                status: 'deployed',
-              },
-            },
-          },
-        ],
-        loadError: null,
-        loaded: true,
-      },
+    helmReleaseDetailsProps = {
       match: {
+        url: '/helm-releases/ns/xyz/release/helm-mysql',
+        isExact: false,
+        path: '/helm-releases/ns/xyz/release/:name',
         params: {
           ns: 'xyz',
         },
-        url: '/helm-releases/ns/xyz/release/helm-mysql',
-        isExact: true,
-        path: '/helm-releases/ns/xyz/release/:name',
       },
     };
-    loadedHelmReleaseDetailsPageProps = {
-      ...helmReleaseDetailsPageProps,
-      helmReleaseData: mockHelmReleases[0],
-    };
+
+    helmReleaseDetailsPage = shallow(<HelmReleaseDetailsPage {...helmReleaseDetailsProps} />);
   });
-  it('should show the loading box if helm release data is not loaded', () => {
-    loadedHelmReleaseDetailsPageProps.helmReleaseData = null;
-    const helmReleaseDetailsPage = shallow(
-      <LoadedHelmReleaseDetailsPage {...loadedHelmReleaseDetailsPageProps} />,
-    );
-    expect(helmReleaseDetailsPage.find(LoadingBox).exists()).toBe(true);
+
+  it('should render the NamespaceBar component', () => {
+    expect(helmReleaseDetailsPage.find(NamespacedPage).exists()).toBe(true);
   });
-  it('should show the loading box if secret is not loaded', () => {
-    loadedHelmReleaseDetailsPageProps.secret.loaded = false;
-    loadedHelmReleaseDetailsPageProps.secret.loadError = undefined;
-    const helmReleaseDetailsPage = shallow(
-      <LoadedHelmReleaseDetailsPage {...loadedHelmReleaseDetailsPageProps} />,
-    );
-    expect(helmReleaseDetailsPage.find(LoadingBox).exists()).toBe(true);
-  });
-  it('should show the status box if there is an error loading the secret', () => {
-    loadedHelmReleaseDetailsPageProps.secret.loadError = 'error 404';
-    const helmReleaseDetailsPage = shallow(
-      <LoadedHelmReleaseDetailsPage {...loadedHelmReleaseDetailsPageProps} />,
-    );
-    expect(helmReleaseDetailsPage.find(StatusBox).exists()).toBe(true);
-  });
-  it('should render the DetailsPage component when secret gets loaded', () => {
-    const helmReleaseDetailsPage = shallow(
-      <LoadedHelmReleaseDetailsPage {...loadedHelmReleaseDetailsPageProps} />,
-    );
-    expect(helmReleaseDetailsPage.find(DetailsPage).exists()).toBe(true);
-  });
-  it('should show the ErrorPage404 for an incorrect release name in the url', () => {
-    loadedHelmReleaseDetailsPageProps.secret.data = [];
-    const helmReleaseDetailsPage = shallow(
-      <LoadedHelmReleaseDetailsPage {...loadedHelmReleaseDetailsPageProps} />,
-    );
-    expect(helmReleaseDetailsPage.find(ErrorPage404).exists()).toBe(true);
+  it('should render the HelmReleaseDetails component', () => {
+    expect(helmReleaseDetailsPage.find(HelmReleaseDetails).exists()).toBe(true);
   });
 });
