+++ frontend/packages/dev-console/src/components/helm/details/__tests__/HelmReleaseNotes.spec.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
 import { SyncMarkdownView } from '@console/internal/components/markdown-view';
+import { mockHelmReleases } from '../../__tests__/helm-release-mock-data';
 import HelmReleaseNotes from '../HelmReleaseNotes';
-import { mockHelmReleases } from './helm-release-mock-data';
 
 describe('HelmReleaseNotes', () => {
   it('should render the SyncMarkdownView component when notes are available', () => {
