+++ frontend/packages/dev-console/src/components/helm/list/HelmReleaseList.tsx
@@ -2,16 +2,16 @@ import * as React from 'react';
 import { FirehoseResult } from '@console/internal/components/utils';
 import { useDeepCompareMemoize } from '@console/shared';
 import { SortByDirection } from '@patternfly/react-table';
-import CustomResourceList from '../custom-resource-list/CustomResourceList';
+import CustomResourceList from '../../custom-resource-list/CustomResourceList';
 import {
-  helmRowFilters,
-  getFilteredItemsByRow,
-  getFilteredItemsByText,
+  helmReleasesRowFilters,
+  filterHelmReleasesByName,
+  filterHelmReleasesByStatus,
   fetchHelmReleases,
-} from './helm-utils';
-import HelmReleaseRow from './HelmReleaseRow';
-import HelmReleaseHeader from './HelmReleaseHeader';
-import { HelmRelease } from './helm-types';
+} from '../helm-utils';
+import HelmReleaseListRow from './HelmReleaseListRow';
+import HelmReleaseListHeader from './HelmReleaseListHeader';
+import { HelmRelease } from '../helm-types';
 
 interface HelmReleaseListProps {
   namespace: string;
@@ -30,13 +30,13 @@ const HelmReleaseList: React.FC<HelmReleaseListProps> = ({ namespace, secrets })
       fetchCustomResources={getHelmReleases}
       dependentResource={memoizedSecrets}
       queryArg="rowFilter-helm-release-status"
-      rowFilters={helmRowFilters}
+      rowFilters={helmReleasesRowFilters}
       sortBy="name"
       sortOrder={SortByDirection.asc}
-      rowFilterReducer={getFilteredItemsByRow}
-      textFilterReducer={getFilteredItemsByText}
-      resourceRow={HelmReleaseRow}
-      resourceHeader={HelmReleaseHeader}
+      rowFilterReducer={filterHelmReleasesByStatus}
+      textFilterReducer={filterHelmReleasesByName}
+      resourceRow={HelmReleaseListRow}
+      resourceHeader={HelmReleaseListHeader}
     />
   );
 };
