+++ frontend/packages/dev-console/src/components/helm/list/HelmReleaseListRow.tsx
@@ -4,11 +4,11 @@ import { Status } from '@console/shared';
 import { TableRow, TableData, RowFunction } from '@console/internal/components/factory';
 import { Timestamp, Kebab, ResourceIcon } from '@console/internal/components/utils';
 import { Link } from 'react-router-dom';
-import { HelmRelease } from './helm-types';
-import { tableColumnClasses } from './HelmReleaseHeader';
-import { deleteHelmRelease, upgradeHelmRelease } from '../../actions/modify-helm-release';
+import { HelmRelease } from '../helm-types';
+import { tableColumnClasses } from './HelmReleaseListHeader';
+import { deleteHelmRelease, upgradeHelmRelease } from '../../../actions/modify-helm-release';
 
-const HelmReleaseRow: RowFunction<HelmRelease> = ({ obj, index, key, style }) => {
+const HelmReleaseListRow: RowFunction<HelmRelease> = ({ obj, index, key, style }) => {
   const menuActions = [
     upgradeHelmRelease(obj.name, obj.namespace),
     deleteHelmRelease(obj.name, obj.namespace),
@@ -46,4 +46,4 @@ const HelmReleaseRow: RowFunction<HelmRelease> = ({ obj, index, key, style }) =>
   );
 };
 
-export default HelmReleaseRow;
+export default HelmReleaseListRow;
